Feature/character training

+++ bootstrap.sh
@@ -11,17 +11,20 @@ mkdir -p windows/false
 
 echo "download necessary data files"
 
-curl "http://www.iapr-tc11.org/dataset/ICDAR2003_RobustReading/TrialTrain/word.zip" -o "word.zip"
-unzip word.zip
-rm word.zip
-clear
+if [ ! -f "word.zip" ]
+then
+  curl "http://www.iapr-tc11.org/dataset/ICDAR2003_RobustReading/TrialTrain/word.zip" -o "word.zip"
+  unzip word.zip
+fi
 
-curl "http://www.iapr-tc11.org/dataset/ICDAR2003_RobustReading/TrialTrain/scene.zip" -o "scene.zip"
-unzip scene.zip
-rm scene.zip
-clear
+if [ ! -f "scene.zip" ]
+then
+  curl "http://www.iapr-tc11.org/dataset/ICDAR2003_RobustReading/TrialTrain/scene.zip" -o "scene.zip"
+  unzip scene.zip
+fi
 
 
+# character recognition
 if [ ! -f character_icdar_train.zip ]; then
   wget -O character_icdar_train.zip http://www.iapr-tc11.org/dataset/ICDAR2003_RobustReading/TrialTrain/char.zip
   wget -O character_icdar_test.zip http://www.iapr-tc11.org/dataset/ICDAR2003_RobustReading/TrialTest/char.zip
@@ -31,7 +34,10 @@ if [ ! -f character_icdar_train.zip ]; then
   unzip -d character_icdar_test character_icdar_test.zip
 fi
 
+
 cd ../src
+
+python3 character_training.py
 
 echo "run training module"
 python3 main.py
